rename capn to devnets (harry hates fun)
