+++ devnets/mod.ts
@@ -2,10 +2,10 @@
 
 export * from "./binary.ts"
 export * from "./CapiConfig.ts"
-export * from "./CapnConnection.ts"
 export * from "./capnHandler.ts"
 export * from "./chainSpec.ts"
 export * from "./createTempDir.ts"
+export * from "./DevnetConnection.ts"
 export * from "./startNetwork.ts"
 export * from "./syncConfig.ts"
 export * from "./testUsers.ts"
