+++ mod.ts
@@ -3,9 +3,8 @@ export { BitSequence } from "./deps/scale.ts"
 
 // moderate --exclude main.ts server util capi.config.ts
 
-export * from "./capi.config.ts"
-export * from "./capn/mod.ts"
 export * from "./crypto/mod.ts"
+export * from "./devnets/mod.ts"
 export * from "./fluent/mod.ts"
 export * from "./frame_metadata/mod.ts"
 export * from "./rpc/mod.ts"
